refactor(dev): Changed 'customTheme' composable to more fluent 'theme'
